ADDED : Method To Load US Census Data And Return Count

+++ src/main/java/com/bridgelabz/censusanalyser/IndiaCensusCSV.java
@@ -10,8 +10,8 @@ public class IndiaCensusCSV {
     public long population;
 
     @CsvBindByName(column = "AreaInSqKm", required = true)
-    public long areaInSqKm;
+    public double areaInSqKm;
 
     @CsvBindByName(column = "DensityPerSqKm", required = true)
-    public int densityPerSqKm;
+    public double densityPerSqKm;
 }
